Documents/Project for Internship work/requirements.txt
torch>=2.0.0
transformers>=4.30.0
snntorch>=0.5.0
faiss-cpu>=1.7.4
numpy>=1.21.0
pytest>=7.0.0
